+++ securedrop/source.py
@@ -429,16 +429,6 @@ def metadata():
     return resp
 
 
-@app.errorhandler(404)
-def page_not_found(error):
-    return render_template('notfound.html'), 404
-
-
-@app.errorhandler(500)
-def internal_error(error):
-    return render_template('error.html'), 500
-
-
 if __name__ == "__main__":  # pragma: no cover
     debug = getattr(config, 'env', 'prod') != 'prod'
     app.run(debug=debug, host='0.0.0.0', port=8080)
